+++ lib/src/HttpControllerBinder.h
@@ -7,6 +7,16 @@
 namespace drogon
 {
 
+class HttpSimpleControllerBinder : public ControllerBinderBase
+{
+  public:
+    void handleRequest(
+        const HttpRequestImplPtr &req,
+        std::function<void(const HttpResponsePtr &)> &&callback) const override;
+
+    std::shared_ptr<HttpSimpleControllerBase> controller_;
+};
+
 class HttpControllerBinder : public ControllerBinderBase
 {
   public:
